apps/api/requirements.txt
fastapi==0.135.2
uvicorn==0.42.0
SQLAlchemy==2.0.48
psycopg==3.3.3
PyMySQL==1.1.1
pydantic==2.12.5
pydantic-settings==2.13.1
PyYAML==6.0.3
